+++ daemon/src/main/java/com/github/andock/daemon/containers/ContainerLogKey.kt
@@ -2,6 +2,6 @@ package com.github.andock.daemon.containers
 
 data class ContainerLogKey(
     val containerId: String,
-    val currentPage: Long,
-    val pageSize: Long,
+    val currentPage: Int,
+    val pageSize: Int,
 )
